refactor: criar classe BaseDatabaseTest para centralizar verificações de banco de dados, reduzindo duplicidade
